@@ -52,12 +52,12 @@ describe "Diffs on operation parents" do
     end
   
     it "has no relevant changes at the beginning" do
-      @nest.op.relevant_changes_for(@queen).should be_empty
+      expect(@nest.op.relevant_changes_for(@queen)).to be_empty
     end
   
     it "includes name in the relevant change" do
       @queen.description['name'] = "Barbara"
-      @op.relevant_changes_for(@queen).should include('queen.description.name')
+      expect(@op.relevant_changes_for(@queen)).to include('queen.description.name')
     end
 
     it "can find relevant changes against any object" do
@@ -65,8 +65,8 @@ describe "Diffs on operation parents" do
       modified = @queen.to_mongo
       modified['description'] = modified['description'].dup
       modified['description']['name'] = "Sue"
-      @nest.op.relevant_changes_for(@queen, :against => original).should be_empty
-      @nest.op.relevant_changes_for(@queen, :against => modified).should_not be_empty
+      expect(@nest.op.relevant_changes_for(@queen, :against => original)).to be_empty
+      expect(@nest.op.relevant_changes_for(@queen, :against => modified)).to_not be_empty
     end
   end
 end
