@@ -56,13 +56,13 @@ describe "MongoPercolator Node & Operation integration" do
     end
 
     it "is frozen" do
-      RealOp.parent_labels.frozen?.should be_true
+      expect(RealOp.parent_labels.frozen?).to be true
     end
 
     it "persists the parent on assignment" do
       @op.animals = @animals
-      @animals.persisted?.should be_true
-      AnimalsIntegration.where(:id => @animals.id).count.should == 1
+      expect(@animals.persisted?).to be true
+      expect(AnimalsIntegration.where(:id => @animals.id).count).to eq(1)
     end
 
     it "doesn't leave any traces if it fails in emit" do
@@ -76,15 +76,15 @@ describe "MongoPercolator Node & Operation integration" do
 
     it "can access the parent using the reader" do
       @op.animals = @animals
-      @op.animals.should == @animals
+       expect(@op.animals).to eq(@animals)
     end
 
     it "has a one association" do
-      @op.should respond_to(:node)
+      expect(@op).to respond_to(:node)
     end
 
     it "has the correct parent labels" do
-      RealOp.parent_labels.to_a.should == [:animals]
+       expect(RealOp.parent_labels.to_a).to eq([:animals])
     end
 
     it "allows operations to be added to another class" do
@@ -92,7 +92,7 @@ describe "MongoPercolator Node & Operation integration" do
       node.create_real_op :animals => AnimalsIntegration.new(:wild => ['baboon'])
       node.real_op.perform!
       node.reload
-      node.pets.should include('baboon')
+      expect(node.pets).to include('baboon')
     end
 
     it "can be performed when the op is created separately" do
@@ -104,13 +104,13 @@ describe "MongoPercolator Node & Operation integration" do
     it "sets up the node when using the create convenience function" do
       node = SomeOtherNode.new
       node.create_real_op :animals => AnimalsIntegration.new(:wild => ['baboon'])
-      node.should_not be_persisted
+      expect(node).to_not be_persisted
       op = node.real_op
-      op.should be_persisted
-      op.should be_naive
-      node.save.should be_true
+      expect(op).to be_persisted
+      expect(op).to be_naive
+      expect(node.save).to be true
       op.reload
-      op.should be_available
+      expect(op).to be_available
     end
 
     it "fails when passed a malformed label" do
@@ -138,19 +138,18 @@ describe "MongoPercolator Node & Operation integration" do
       end
 
       it "saves the parent on assignment" do
-        @animals.should be_persisted
+        expect(@animals).to be_persisted
       end
 
       it "knows which dependencies have changed (1)" do
         @animals.wild << 'meerkat'
-        @op.relevant_changes_for(@animals).should == ['animals.wild']
+         expect(@op.relevant_changes_for(@animals)).to eq(['animals.wild'])
       end
   
       it "knows which dependencies have changed (2)" do
         @animals.wild << 'meerkat'
         @animals.farm = ['pig']
-        @op.relevant_changes_for(@animals).sort.should == 
-          %w(animals.farm animals.wild)
+        expect(@op.relevant_changes_for(@animals).sort).to eq(%w(animals.farm animals.wild))
       end
     end
   end
@@ -159,7 +158,7 @@ describe "MongoPercolator Node & Operation integration" do
     it "can be performed when the op is created separately" do
       op = SomeNodeWithRequirement::Op.new
       node = SomeNodeWithRequirement.new :op => op
-      node.should_not be_persisted
+      expect(node).to_not be_persisted
       node.name = "King George"
       node.op.perform_on! node
     end
@@ -178,7 +177,7 @@ describe "MongoPercolator Node & Operation integration" do
     end
 
     it "has an operation association" do
-      @node.should respond_to(:real_op)
+      expect(@node).to respond_to(:real_op)
     end
     
     context "has an op associated" do
@@ -188,36 +187,36 @@ describe "MongoPercolator Node & Operation integration" do
       end
 
       it "has a real op associated" do
-        @node.real_op.should be_kind_of(RealOp)
+        expect(@node.real_op).to be_kind_of(RealOp)
       end
 
       it "can find the node from the real op" do
-        @op.node.should be_kind_of(SomeNode)
+        expect(@op.node).to be_kind_of(SomeNode)
       end
   
       it "can compute computed properties on demand" do
-        @node.pets.should == []
+         expect(@node.pets).to eq([])
         @op.perform!
         @node.reload
-        @node.pets.should == %w(binturong dugong sloth)
+         expect(@node.pets).to eq(%w(binturong dugong sloth))
       end
 
       it "destroys its operations when node is destroyed" do
         real_op_id = @node.real_op.id
         node_id = @node.id
-        RealOp.find(real_op_id).should_not be_nil
+        expect(RealOp.find(real_op_id)).to_not be_nil
         @node.destroy
-        SomeNode.find(node_id).should be_nil
-        RealOp.find(real_op_id).should be_nil
+        expect(SomeNode.find(node_id)).to be_nil
+        expect(RealOp.find(real_op_id)).to be_nil
       end
 
       it "destroys the old operation when it's replaced" do
         old_id = @node.real_op.id
-        RealOp.find(old_id).should_not be_nil
+        expect(RealOp.find(old_id)).to_not be_nil
         new_op = RealOp.new :animals => @animals
         @node.real_op = new_op
-        @node.real_op.id.should_not == old_id
-        RealOp.find(old_id).should be_nil
+         expect(@node.real_op.id).to_not eq(old_id)
+        expect(RealOp.find(old_id)).to be_nil
       end
 
       context "computed initially" do
@@ -227,49 +226,49 @@ describe "MongoPercolator Node & Operation integration" do
         end
 
         it "should be marked as old when the parent is changed" do
-          @node.real_op.stale?.should be_false
+          expect(@node.real_op.stale?).to be false
           @animals.farm = ['hog']
           @animals.save!
           @node.reload
-          @node.real_op.stale?.should be_true
+          expect(@node.real_op.stale?).to be true
         end
 
         it "should be marked as old when the identity of the parent changes" do
-          @node.real_op.stale?.should be_false
+          expect(@node.real_op.stale?).to be false
           new_animals = AnimalsIntegration.create :farm => ['sheep']
-          new_animals.should_not be_nil
+          expect(new_animals).to_not be_nil
           @node.real_op.animals = new_animals
-          @node.real_op.save.should be_true
-          @node.real_op.stale?.should be_true
+          expect(@node.real_op.save).to be true
+          expect(@node.real_op.stale?).to be true
         end
     
         it "gets an updated computed property when the parent is changed" do
           @animals.farm = ['hog']
           @animals.save
-          MongoPercolator::Operation.where(:stale => true).count.should == 1
-          MongoPercolator.percolate.operations.should == 1
+          expect(MongoPercolator::Operation.where(:stale => true).count).to eq(1)
+          expect(MongoPercolator.percolate.operations).to eq(1)
           @node.reload
-          @node.pets.should == %w(binturong dugong hog sloth)
+           expect(@node.pets).to eq(%w(binturong dugong hog sloth))
         end
 
         it "is no longer marked as old after percolation" do
           @animals.farm = ['hog']
-          @animals.save.should be_true
-          MongoPercolator.percolate.operations.should == 1
+          expect(@animals.save).to be true
+          expect(MongoPercolator.percolate.operations).to eq(1)
           @node.reload
-          @node.real_op.stale?.should be_false
+          expect(@node.real_op.stale?).to be false
         end
 
         it "only takes the number of passes needed" do
           @animals.farm = ['hog']
           @animals.save
-          MongoPercolator.percolate.operations.should == 1
+          expect(MongoPercolator.percolate.operations).to eq(1)
         end
 
         it "can use the guide to percolate" do
           @animals.farm = ['hog']
           @animals.save
-          MongoPercolator.guide.percolate.operations.should == 1
+          expect(MongoPercolator.guide.percolate.operations).to eq(1)
         end
 
         it "uses the cache when computing dependencies" do
@@ -278,7 +277,7 @@ describe "MongoPercolator Node & Operation integration" do
           node2.real_op = op2
           op2.save!
           node2.save!
-          RealOp.should_receive(:relevant_changes_for).once.and_call_original
+          expect(RealOp).to receive(:relevant_changes_for).once.and_call_original
           @animals.save!
         end
       end
