@@ -9,25 +9,25 @@ describe MongoPercolator::Many::Copy do
     end
   end
 
-  it { should respond_to(:ids) }
-  it { should respond_to(:root) }
-  it { should respond_to(:root_id) }
-  it { should respond_to(:path) }
-  it { should respond_to(:label) }
+  it { expect(subject).to respond_to(:ids) }
+  it { expect(subject).to respond_to(:root) }
+  it { expect(subject).to respond_to(:root_id) }
+  it { expect(subject).to respond_to(:path) }
+  it { expect(subject).to respond_to(:label) }
 
   it "can be saved with factory defaults" do
-    copy.save.should be_true
+    expect(copy.save).to be true
   end
 
   it "fails to be saved without a label" do
     copy.label = nil
-    copy.save.should be_false
-    copy.errors[:label].should_not be_nil
+    expect(copy.save).to be false
+    expect(copy.errors[:label]).to_not be_nil
   end
 
   it "fails to be saved without a root" do
     copy.root_id = nil
-    copy.save.should be_false
-    copy.errors[:root].should_not be_nil
+    expect(copy.save).to be false
+    expect(copy.errors[:root]).to_not be_nil
   end
 end
