@@ -1,30 +1,27 @@
 require 'spec_helper.rb'
 
 describe MongoPercolator do
-  it { should respond_to(:whoami?) }
+  it { expect(subject).to respond_to(:whoami?) }
 
   it "deletes _id" do
-    subject.dup_hash_selectively({'_id' => 1, 'other' => 2}).
-      should == {'other' => 2}
+    expect(subject.dup_hash_selectively({'_id' => 1, 'other' => 2})).to eq({'other' => 2})
   end
 
   it "deletes updated_at" do
-    subject.dup_hash_selectively({'updated_at' => 1, 'other' => 2}).
-      should == {'other' => 2}
+    expect(subject.dup_hash_selectively({'updated_at' => 1, 'other' => 2})).to eq({'other' => 2})
   end
 
   it "deletes created_at" do
-    subject.dup_hash_selectively({'created_at' => 1, 'other' => 2}).
-      should == {'other' => 2}
+    expect(subject.dup_hash_selectively({'created_at' => 1, 'other' => 2})).to eq({'other' => 2})
   end
 
   it "can duplicate a one-level hash without ids" do
     hash = {'_id' => "fresh air", 'dave' => 'davies', 'terry' => 'gross'}
     hash_out = subject.dup_hash_selectively(hash)
-    hash_out.should_not include('_id')
-    hash.should include('_id')
-    hash_out['dave'].should == 'davies'
-    hash_out['terry'].should == 'gross'
+    expect(hash_out).to_not include('_id')
+    expect(hash).to include('_id')
+    expect(hash_out['dave']).to eq('davies')
+    expect(hash_out['terry']).to eq('gross')
   end
 
   it "can recursively duplicate a two-level hash involving an array" do
@@ -34,14 +31,14 @@ describe MongoPercolator do
 
     # Check the output hash
     hash_out = subject.dup_hash_selectively(hash)
-    hash_out.should_not include('_id')
-    hash_out['people'][0].should == {'dave' => 'davies'}
-    hash_out['people'][1].should == {'terry' => 'gross'}
-    hash_out.length.should == 1
+    expect(hash_out).to_not include('_id')
+    expect(hash_out['people'][0]).to eq({'dave' => 'davies'})
+    expect(hash_out['people'][1]).to eq({'terry' => 'gross'})
+    expect(hash_out.length).to eq(1)
 
     # Check the original hash
-    hash['_id'].should == 'fresh air'
-    hash['people'][0].should == {'_id' => 'stand in', 'dave' => 'davies'}
-    hash['people'][1].should == {'_id' => 'host', 'terry' => 'gross'}
+    expect(hash['_id']).to eq('fresh air')
+    expect(hash['people'][0]).to eq({'_id' => 'stand in', 'dave' => 'davies'})
+    expect(hash['people'][1]).to eq({'_id' => 'host', 'terry' => 'gross'})
   end
 end
