@@ -22,26 +22,26 @@ describe "Sci Hacks" do
       SciHackTest2.where(:name => "Bob").upsert :$set => {:x => 33}, 
         :$inc => {:z => 88, :y => 9}
       doc = SciHackTest2.first
-      doc.name.should == "Bob"
-      doc.x.should == 33
-      doc.y.should == 9
-      doc.z.should == 88
+       expect(doc.name).to eq("Bob")
+      expect(doc.x).to eq(33)
+      expect(doc.y).to eq(9)
+      expect(doc.z).to eq(88)
     end
   
     it "Can create a new document without sci" do
       SciHackTest.where(:name => "Bob").upsert :$set => {:x => 33}, 
         :$inc => {:y => 9}
       doc = SciHackTest.first
-      doc.should be_a(SciHackTest)
-      doc.name.should == "Bob"
-      doc.x.should == 33
-      doc.y.should == 9
+      expect(doc).to be_a(SciHackTest)
+      expect(doc.name).to eq("Bob")
+      expect(doc.x).to eq(33)
+      expect(doc.y).to eq(9)
     end
 
     it "retuns the node from the upsert" do
       doc = SciHackTest.where(:name => "Major Major").upsert :$set => {:x => 22}
-      SciHackTest.count.should == 1
-      doc.id.should == SciHackTest.first.id
+      expect(SciHackTest.count).to eq(1)
+       expect(doc.id).to eq(SciHackTest.first.id)
     end
   end
 end
